feat: add VersionUpdateModifier
