+++ CodeLounge.xcodeproj/xcuserdata/kimdonghyeon.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -28,7 +28,15 @@
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>DFD31FD72D36175F0076DD16</key>
+		<dict>
+			<key>primary</key>
+			<true/>
 		</dict>
 	</dict>
 </dict>
