--- CodeLounge/ContentView.swift
@@ -1,18 +1,0 @@
-//
-//  ContentView.swift
-//  CodeLounge
-//
-//  Created by 김동현 on 1/14/25.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        IntroView()
-    }
-}
-
-#Preview {
-    ContentView()
-}
