+++ CodeLounge/General/Constant.swift
@@ -2,17 +2,20 @@
 //  Constant.swift
 //  CodeLounge
 //
-//  Created by 김동현 on 1/16/25.
+//  Created by 김동현 on 4/3/26.
 //
 
 import Foundation
 
 enum Constant {}
-
 extension Constant {
-    struct DBKey {
-        static let Users = "Users"
-    }
+  enum DBKey {
+    static let Users = "Users"
+  }
+  
+  enum URL {
+    static let appStore = "https://itunes.apple.com/lookup?bundleId=com.indextrown.CodeLounge"
+  }
 }
 
-typealias DBKey = Constant.DBKey
+// typealias DBKey = Constant.DBKey
