--- CodeLounge/Service/Services.swift
@@ -1,28 +1,0 @@
-//
-//  Services.swift
-//  CodeLounge
-//
-//  Created by 김동현 on 1/14/25.
-//
-
-import Foundation
-
-protocol ServiceType {
-    var authService: AuthenticationServiceType { get set }
-    var userService: UserServiceType { get set }
-}
-
-final class Services: ServiceType {
-    var authService: AuthenticationServiceType
-    var userService: UserServiceType
-    
-    init() {
-        self.authService = AuthenticationService()
-        self.userService = UserService(dbRepository: UserDBRepository())
-    }
-}
-
-final class StubServices: ServiceType {
-    var authService: AuthenticationServiceType = StubAuthenticationService()
-    var userService: UserServiceType = UserService(dbRepository: UserDBRepository())
-}
