--- CodeLounge/View/Authentication/AuthenticationView.swift
@@ -1,35 +1,0 @@
-//
-//  AuthenticationView.swift
-//  CodeLounge
-//
-//  Created by 김동현 on 1/16/25.
-//
-
-import SwiftUI
-
-struct AuthenticationView: View {
-    @StateObject var authViewModel: AuthenticationViewModel
-    
-    var body: some View {
-        VStack {
-            switch authViewModel.authenticationState {
-            case .unauthenticated:
-                IntroView()
-                    .environmentObject(authViewModel)
-            case .authenticated:
-                MainTabView()
-                    .environmentObject(authViewModel)
-            case .firstTimeLogin:
-                NicknameSettingView()
-                    .environmentObject(authViewModel)
-            }
-        }
-        .onAppear {
-            authViewModel.send(action: .checkAuthenticationState)
-        }
-    }
-}
-
-#Preview {
-    AuthenticationView(authViewModel: AuthenticationViewModel(container: DIContainer(services: StubServices())))
-}
